separate testutils from gemmini.h
