+++ bareMetalC/tiled_matmul_cpu.c
@@ -8,7 +8,7 @@
 #ifndef BAREMETAL
 #include <sys/mman.h>
 #endif
-#include "include/gemmini.h"
+#include "include/gemmini_testutils.h"
 
 #define CHECK_RESULT 1
