+++ bareMetalC/tiled_matmul_option.c
@@ -8,7 +8,7 @@
 #ifndef BAREMETAL
 #include <sys/mman.h>
 #endif
-#include "include/gemmini.h"
+#include "include/gemmini_testutils.h"
 
 #ifndef BAREMETAL
 #define MAT_DIM_I 256
